define wsapi_checks in terms of cru_checks
Like the previous commit, we'll use xml diffs
to maintain (legacy) project-specific variations from cru_checks.

Note: I thought about putting this in wsapi itself,
but this approach is less complex.
I'm not yet totally convinced this is the better approach,
but I think it's the best way to start.